requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0

# Training and logging
wandb>=0.16.0
tensorboard>=2.14.0

# Data and utilities
pyyaml>=6.0
jsonschema>=4.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Optional (for development)
ipython>=8.12.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
